feat(tools): move WebhookCapability POD to dasclaw_tool (#672 step 2c)
Verbatim-port the 7-field `WebhookCapability` POD struct from
`desktop-client/ironclaw/src/tools/wasm/capabilities.rs` to a new
`crates/dasclaw_tool/src/webhook_capability.rs`, and switch
`Tool::webhook_capability()` + `WasmToolWrapper::webhook_capability()`
to return `Option<dasclaw_tool::WebhookCapability>`. ironclaw side keeps
`pub use dasclaw_tool::WebhookCapability` so every existing call site
(`Capabilities.webhook`, `webhooks/mod.rs` test stubs, capabilities_schema)
keeps compiling unchanged.

Per ADR-154 §3.5 — closes the second of the two #672 blockers (the first,
JobContext coupling, was resolved by #682 step 2b). With both blockers
gone, the next slice (#672 step 3) can physically relocate the `Tool`
trait to `dasclaw_tool`.

## Sources read
- AGENTS.md (Skills 强制使用规范, PR/commit discipline)
- .github/copilot-instructions.md (§3 mandatory pipeline, §4 PR checklist, §6 reuse-before-create)
- docs/plans/architecture-refactor/adr-154-jobcontext-core-trait-split.md (§3.5 WebhookCapability 同步迁移, §4 non-goals)
- desktop-client/ironclaw/src/tools/wasm/capabilities.rs (current struct location, verbatim source)
- desktop-client/ironclaw/src/tools/tool.rs (trait default method)
- desktop-client/ironclaw/src/tools/wasm/wrapper.rs (impl that returns it)
- desktop-client/ironclaw/src/webhooks/mod.rs (test stubs that construct it)

## Verification
```
cargo check -p dasclaw_tool --tests   # OK
cargo check -p dasclaw --tests        # OK (warning-free after removing dead re-export)
cargo fmt --all -- --check            # clean
python3.12 scripts/check_no_panics.py --base origin/xClaw   # OK
cargo clippy --no-deps -p dasclaw_tool --all-targets -- -D warnings   # clean
cargo clippy --no-deps -p dasclaw --all-targets -- -D warnings        # clean (5m12s)
```

Closes #684
Refs #672
Refs #683
